Split tendermint websocket client into multiple files
