+++ packages/iov-tendermint-rpc/types/rpcclients/utils.d.ts
@@ -1,0 +1,4 @@
+import { JsonRpcResponse } from "@iov/jsonrpc";
+import { SocketWrapperMessageEvent } from "@iov/socket";
+export declare function defaultErrorHandler(error: any): never;
+export declare function toJsonRpcResponse(message: SocketWrapperMessageEvent): JsonRpcResponse;
